Update candidString parameter in createActorContext and useActor
