+++ packages/react/src/context/actor/types.ts
@@ -1,0 +1,30 @@
+import type {
+  IDL,
+  ActorHooksReturnType,
+  BaseActor,
+  ActorManagerParameters,
+  CanisterId,
+} from "../../types"
+
+export interface CreateActorContextReturnType<A = BaseActor>
+  extends ActorHooksReturnType<A> {
+  ActorProvider: React.FC<ActorProviderProps>
+}
+
+export interface ActorProviderProps extends CreateActorContextParameters {
+  children?: React.ReactNode | undefined
+  candidString?: string
+  loadingComponent?: React.ReactNode
+  authenticatingComponent?: React.ReactNode
+}
+
+export interface CreateActorContextParameters
+  extends Omit<
+    ActorManagerParameters,
+    "idlFactory" | "agentManager" | "canisterId"
+  > {
+  didjsId?: string
+  canisterId?: CanisterId
+  idlFactory?: IDL.InterfaceFactory
+  initializeParser?: boolean
+}
