+++ packages/react/src/hooks/adapter/index.ts
@@ -1,0 +1,1 @@
+export { useCandidAdapter } from "./useCandidAdapter"
