+++ packages/react/src/hooks/agent/index.ts
@@ -4,4 +4,4 @@ export * from "./useAgentManager"
 export * from "./useAuth"
 export * from "./useAuthState"
 export * from "./useUserPrincipal"
-export * from "./useCandidAdapter"
+export * from "../adapter/useCandidAdapter"
