--- packages/react/src/hooks/agent/useCandidAdapter.ts
@@ -1,61 +1,0 @@
-import { useCallback, useEffect, useState } from "react"
-import { createCandidAdapter } from "@ic-reactor/core"
-import { useAgent } from "./useAgent"
-
-import type { CandidAdapter } from "@ic-reactor/core/dist/classes"
-export type { CandidAdapter }
-
-export interface UseCandidAdapterParams {
-  didjsCanisterId?: string
-  initializeParser?: boolean
-}
-
-/**
- * Accesses the `CandidAdapter` to download the actor's Candid interface.
- *
- * @param config - `UseCandidAdapterParams` The configuration object.
- * @returns The `CandidAdapter` instance.
- *
- * @example
- * ```jsx
- * function CandidAdapterComponent() {
- *   const candidAdapter = useCandidAdapter();
- *
- *   const getActor = async () => {
- *      const { idlFactory } = await candidAdapter.getCandidDefinition(canisterId)
- *      console.log(idlFactory)
- *   }
- *
- *   return (
- *       <button onClick={getActor}>Get Actor</button>
- *   );
- * }
- *```
- */
-export const useCandidAdapter = (config: UseCandidAdapterParams = {}) => {
-  const [candidAdapter, setCandidAdapter] = useState<CandidAdapter>()
-
-  const agent = useAgent()
-
-  const initializeCandidAdapter = useCallback(async () => {
-    try {
-      const candidAdapter = createCandidAdapter({
-        agent,
-        didjsCanisterId: config.didjsCanisterId,
-      })
-      if (config.initializeParser) {
-        await candidAdapter.initializeParser()
-      }
-      setCandidAdapter(candidAdapter)
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.error("Error creating CandidAdapter", error)
-    }
-  }, [agent])
-
-  useEffect(() => {
-    initializeCandidAdapter()
-  }, [initializeCandidAdapter])
-
-  return candidAdapter
-}
