+++ packages/react/src/types.ts
@@ -19,7 +19,7 @@ export interface CreateReactorReturnType<A>
   getVisitFunction: () => VisitService<A>
 }
 
-export * from "./context/types"
+export * from "./context/agent/types"
 export * from "./helpers/types"
 export * from "./hooks/types"
